fix wrong words and ordering for staff member form
