+++ src/staff/components/UserStatus/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./UserStatus";
+export * from "./UserStatus";
